Bug#37804715: INTERSECT operation return wrong result
During computation for INTERSECT and EXCEPT, we may encounter a
situation where we try to read a row containing a NULL field into the
temporary result table which has this field in the result set as not
nullable. For EXCEPT, this can happen when the left operand has this
field as non-nullable. For INTERSECT it can happen if any of the
operands has this field as non-nullable, cf. nullability computation
in Query_term_set_op::prepare_query_term, and one of the other
operands allow NULLs in this field (position).

This is detected in set_field_to_null_with_conversions, which will
return an error condition: TYPE_ERR_NULL_CONSTRAINT_VIOLATION.

Unfortunately, this was not propagated and bad things happened.

Solution: propagate this error back to the set operation execution
logic, and skip such rows, since they cannot contribute to the result.
For non-aggregated expressions, a filter predicate is added to the
WHERE clause. For aggregated expressions, the filter predicate is added
to the HAVING clause and for windowed expressions, the filter predicate
is added to the QUANTIFY clause.

Test main.query_expressions-bug required adjustments due to this change.
The added filter meant that number of rows selected from the query
blocks of the set operation was reduced, thus invalidating the test
case's purpose. The test case had to be changed so that the expression

  "can't" OR 447938560

was replaced with an equivalent but nullable expression.
Without this change, a filter predicate was added to the other query
block of the INTERSECT operation, which filtered out enough rows to make
the test meaningless.

A few test cases from duplicate bug reports have also been added to
main.query_expression-bugs.

A test change was also recorded for a couple Rapid tests, as some
INTERSECT queries that used to be rejected are now accepted.

Change-Id: Ie3b2ff469b97aed37d6d1f92a7f31ba31d625da0