Bug#38814585  Test failure: innodb.innodb_dirty_pages_at_shutdown
Symptom:
  The test innodb.innodb_dirty_pages_at_shutdown fails sporadically with
MTR reporting 'Command "shutdown_server" failed with error 2'

Root Cause:
  The error 2 indicates that the shutdown command timed out after
waiting for 60 seconds. The test expects the server to crash during the
last phase of shutdown. A core file is generated when the server
crashes. The timeout is hit when shutdown takes too long due to
generation of core file.

Fix:
  Disable the generation of core file as it is not required for the
test.

Change-Id: I37f9777b97f564b49cfddfa7271258f974df6053