Bug#38704535: Assertion `std::cmp_less(pred_idx, m_graph->num_where_predicates)' failed.
An assertion failed in ApplyDelayedPredicatesAfterJoin() during
optimization of a query when using the hypergraph optimizer.

The reason for the assertion was that a sargable join predicate had
been pushed down on both sides of a join. This confused the logic in
ApplyDelayedPredicatesAfterJoin(), which assumed that a sargable join
predicate would only be pushed down to one side of the join; the inner
side of a nested loop join.

This happened because one of the tables, a materialized derived table
with a predicate that always evaluated to false, had a join condition
pushed down to it. The REF access path created for it had the
parameter_tables property correctly set to include the other table
referenced in the join predicate, and this would usually force the REF
access path to the inner side of the nested loop join. However,
because the derived table had a predicate that was known to always
evaluate to false, the REF access path got replaced by a ZERO_ROWS
access path. The bitsets for applied and subsumed sargable join
predicates were populated in the ZERO_ROWS access path, but not the
parameter_tables bitmap. This inconsistency was detected by the
assertion in ApplyDelayedPredicatesAfterJoin().

It is fixed by calling the existing
CostingReceiver::ProposeZeroRowsAccessPath() helper function for
pruning a single table from the join tree. It ensures that the bitsets
of the proposed access path are consistent.

Also, the logic to check if the entire join could be optimized away
because of the empty table, was moved to ProposeZeroRowsAccessPath(),
so that it is now checked in a single place.

Change-Id: Ib3150b3a9dec095bad11c804cc55b6e37a48fb63