Bug#35414680 fil_update_partition_name not needed to be called in fil_tablespace_redo_create
The fil_update_partition_name was introduced in 2019 to handle upgrade
scenario by making sure (sub)partition names and separators used in the
suffix of the filename and the logical space name are written using
lower-case for consistency.

As we are several years and upgrades after this was needed, this code
never does anything, because all the names generated nowadays follow
this new convention.

Frankly, the way this function was implemented and used was dubious
anyway, as callers often ignored the output arguments, which seems
to defeat the whole purpose of calling it.

This patch removes this function and its usages.

Change-Id: I40f513d2be2f00c0d53ab80904de9d057b2ab279