Bug#36711659 lock_remove_recovered_trx_record_locks is not needed since 2013
The lock_remove_recovered_trx_record_locks was meant to find the
record locks acquired via implicit-to-explicit conversion on
behalf of recovered transactions on a specific table, which couldn't
be found by an existing and more general mechanism that iterates
over all the transactions which have a lock on the table.

This additional mechanism was introduced in 2010, as at that time
we have not been resurrecting table locks for the recovered transactions
and thus they indeed couldn't be found.

But, since 2013, a recovered transaction which has modified a table
would have an IX TABLE lock resurrected on its behalf, and thus the
general purpose mechanism would find it and is sufficient.
Therefore, since then the lock_remove_recovered_trx_record_locks is
basically a no-op as confirmed by test coverage information.

This patch removes this function.

Change-Id: I17a6df2a96a78f75720b4a68f0109336201a8a4b