BUG#38609554: Relax GR member join version compatibility rules on 9.7 LTS
With the release model supporting LTS and innovation releases,
inplace downgrade must be supported, but only within the same LTS
release series.
For server versions 9.7.0+ (9.7.1, 9.7.2, etc.) inplace downgrade
shall be allowed.

The same reasoning applies to cross-version groups, that is, a
server 9.7.0+ should be allowed to join a 9.7.1 or higher 9.7
patch version.
Example: on a group with S1 (9.7.1) and S2 (9.7.2), a server with
version 9.7.0 is allowed to join.

For this rule to apply all servers version must be 9.7, that is,
if a higher major version server is part of the group it does not
apply.
Example: on a group with S1 (9.7.1) and S2 (10.0.0), a server with
version 9.7.0 is not allowed to join.

The change of this rule will not impact the primary election
algorithm.
On a primary failover, the elected primary will be from the lowest
version.
Example:
 1) A group with S1 Primary (9.7.1) and S2 Secondary (9.7.2),
    a server with version 9.7.0 is allowed to join and will become
    S3 Secondary (9.7.0).
 2) If S1 Primary leaves the group, the new primary will be S3
    (9.7.0).

Likewise, only the lowest version servers can be used on
group_replication_set_as_primary().
Example:
 1) A group with S1 Primary (9.7.1) and S2 Secondary (9.7.2),
    a server with version 9.7.0 is allowed to join and will become
    S3 Secondary (9.7.0).
 2) group_replication_set_as_primary() will only accept S3
    (9.7.0) UUID.

This patch considers all versions which:
 * major version is equal or greater than 9
 * minor version is equal to 7
a LTS version.

Change-Id: I1b6b774e79cb7e96ad3647dbfd8bf466db81c4ad