Bug#38711166 - Allow dropping database even when database directory is missing.
When dropping a database, the existence of the database is verified
from the data dictionary. If the database exists in the data
dictionary, then the existence of the corresponding directory on
the file system is also checked. If the directory is missing, an
error is reported.

Historically, before the introduction of the transactional data
dictionary, the existence of a database was verified by checking
the corresponding directory on the file system. However, with the
transactional data dictionary, the existence of the database can
be verified using only the data dictionary. Instead of failing the
DROP DATABASE operation due to a missing directory, the operation
now continues and reports a warning about the missing directory.

Furthermore, the following situation due to an inconsistent state
is now avoided:

WITHOUT FIX:
  mysql>CREATE DATABASE db1;

  // Manually removed database directory for db1 from data directory.

  mysql>DROP DATABASE db1;
  ERROR 3679 (HY000): Schema directory './db1/' does not exist

  mysql> CREATE DATABASE db1;
  ERROR 1007 (HY000): Can't create database 'db1'; database exists

In this scenario, the directory for db1 is missing, but the data
dictionary still contains information about db1. As a result, the
db1 database cannot be dropped or re-created.

With the fix, DROP DATABASE executes successfully and information
from the data dictionary is also removed.

Change-Id: I803b4dde41904b3071cdbc43939630a1fe65b60a