Bug#36054459 [4/4] TC node failure handling may still be in progress when node rejoins
New test testNodeRestart -n PreparedUpdatesNF.

Test prepares transactions with many updates. When a node that
participate in the transaction but is not the transaction coordinator
node the transaction will be aborted as part of node fail handling on
the transaction coordinator node. Before fix the node fail handling
could end before transaction was aborted.

Change-Id: I4b60c54ca664dc543ecb71ba01d4c59cb833bd61