Bug#38757640 Ndb binlog purger race causes one stale row
Problem:
There is a race between the NDB binlog injector and the NDB binlog
purger that can leave a stale row in mysql.ndb_binlog_index
for a binlog file that has just been purged.

 Sequence:
 1. The NDB injector session commits the epoch transaction to the
   binlog and determines the binlog file name for that commit.
 2. The NDB injector session then uses a separate transaction to write
   that file name into mysql.ndb_binlog_index (InnoDB).

Because steps (1) and (2) are two different transactions, the purger
can start and complete its removal for that File in between (1) and (2).
In that case, the purger will not see the row inserted in (2), leaving
a single orphan row for the just‑purged file.

Analysis:
The purger should wait until NDB binlog has stopped using the binlog file
which is about to be purged. Since NDB binlog usage of file to be purged
is very rare, the solution needs to be lightweight and only activate
when
purger is requesting the current binlog file name.

Solution:
Extend NDB binlog with functionality to publish the current binlog file
name "in use" when someone is waiting for it.
Make the purger wait until the ndb binlog is not using the binlog file
to be purged.
This patch also extends rpl_injector with functionality to get the
current binlog file name, this effectively encapsulates all acess to
binlog.h which can thus be removed from other place.

Change-Id: Ic70b7ebf1508c39945e1dff8bed5eaaf722512a6